requirements.txt
aiogram~=2.19
environs~=8.0.0
pandas
asyncpg
python-telegram-bot-pagination
xlsxwriter
